# web/packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: Teggo B2B Commerce API
  version: 0.1.0
  description: >
    The single source of truth for the Teggo API contract. The TypeScript
    client used by both frontends is generated from this file (do not hand-edit
    the generated schema). Grown module-by-module as the backend lands.
servers:
  - url: http://localhost:8080
    description: Local Go API
...
components:
  securitySchemes:
    adminAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    storefrontAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    vendorAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
        org_id: { type: integer, format: int64 }
    AuthToken:
      type: object
      required: [token]
      properties:
        token: { type: string }
...
    InviteInfo:
      type: object
      required: [company_name, role, expires_at]
      properties:
        company_name: { type: string }
        role: { type: string }
        expires_at: { type: string, format: date-time }
    InviteAcceptRequest:
      type: object
      required: [email, full_name, password]
      properties:
        email: { type: string, format: email }
        full_name: { type: string }
        password: { type: string, minLength: 8 }
...
    SignupRequest:
      type: object
      required: [organization, full_name, email, password, subdomain]
      properties:
        organization: { type: string, description: 'Company / organization display name.' }
        full_name: { type: string, description: "The first admin user's name." }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        subdomain: { type: string, description: 'Lowercase letters, digits, hyphens (3-63 chars). The storefront lives at <subdomain>.<platform base domain>.' }
        currency: { type: string, description: 'ISO 4217; defaults to USD.' }
        locale: { type: string, description: 'Defaults to en.' }
    SignupResponse:
      type: object
      properties:
        domain: { type: string, description: 'The storefront domain that was provisioned.' }
        message: { type: string }
    SignupVerifyRequest:
      type: object
      required: [token]
      properties:
        token: { type: string, format: uuid }
    MessageResponse:
      type: object
      properties:
        message: { type: string }
    PlatformOrganization:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        status: { type: string, enum: [pending, trial, active, suspended] }
        user_count: { type: integer, format: int64 }
        website_count: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
    ListWrapperPlatformOrganization:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/PlatformOrganization' } }
    OrgStatusRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [trial, active, suspended] }
    OrgStatusResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        status: { type: string }
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
  /admin/auth/login:
    post:
      tags: [Auth]
      operationId: adminLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/AuthToken' } } }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /storefront/auth/login:
    post:
      tags: [Auth]
      operationId: storefrontLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/AuthToken' } } }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /storefront/invites/{token}:
    parameters:
      - { name: token, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Auth]
      operationId: storefrontGetInvite
      summary: Resolve an invite link so the join page can show the company being joined.
      security: []
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/InviteInfo' } } }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '410': { $ref: '#/components/responses/ErrorResponse' }
  /storefront/invites/{token}/accept:
    parameters:
      - { name: token, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Auth]
      operationId: storefrontAcceptInvite
      summary: Self-register a buyer via an invite link and sign them in.
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/InviteAcceptRequest' } } }
      responses:
        '201':
          description: Account created — returns a storefront token (same shape as login).
          content: { application/json: { schema: { $ref: '#/components/schemas/AuthToken' } } }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
        '410': { $ref: '#/components/responses/ErrorResponse' }
...
  /vendor/auth/login:
    post:
      tags: [VendorPortal]
      operationId: vendorLogin
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AuthToken' } } } }
        '401': { $ref: '#/components/responses/ErrorResponse' }
...
  /signup:
    post:
      tags: [Platform]
      operationId: signup
      summary: Self-serve tenant signup — provisions a pending organization and emails a verification link.
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SignupRequest' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/SignupResponse' } } } }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /signup/verify:
    post:
      tags: [Platform]
      operationId: signupVerify
      summary: Consume an emailed verification token; the organization moves to trial and can sign in.
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SignupVerifyRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } }
        '400': { $ref: '#/components/responses/ErrorResponse' }
  /admin/platform/organizations:
    get:
      tags: [Platform]
      operationId: platformListOrganizations
      summary: Platform-operator overview of every tenant organization.
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperPlatformOrganization' } } } }
  /admin/platform/organizations/{id}/status:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    post:
      tags: [Platform]
      operationId: platformSetOrganizationStatus
      summary: Suspend, reactivate, or change a tenant organization's lifecycle status.
      security: [{ adminAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/OrgStatusRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OrgStatusResponse' } } } }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }